Move contents of SiStripLorentzAngleDepESProducer header to source
Also removed unused and commented members
